ISSITES-5632 Rename client extension
