--- workspaces/liferay-learn-workspace/client-extensions/learn-dxp-importer/LCP.json
@@ -1,12 +1,0 @@
-{
-	"cpu": 4,
-	"environments": {
-		"infra": {
-			"deploy": false
-		}
-	},
-	"id": "learndxpimporter",
-	"kind": "Job",
-	"memory": 3000,
-	"scale": 1
-}
