--- workspaces/liferay-learn-workspace/client-extensions/learn-dxp-importer/source-formatter-suppressions.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0"?>
-
-<suppressions>
-	<checkstyle>
-		<suppress checks="ExceptionPrintStackTraceCheck" files=".*" />
-	</checkstyle>
-	<source-check>
-		<suppress checks="GradleDependencyArtifactsCheck" files=".*" />
-	</source-check>
-</suppressions>
