--- workspaces/liferay-learn-workspace/client-extensions/learn-dxp-importer/source-formatter.properties
@@ -1,9 +1,0 @@
-##
-## Exclusions
-##
-## See https://github.com/liferay/liferay-portal/blob/master/source-formatter.properties
-## for available properties.
-##
-
-    checkstyle.ChainingCheck.allowedClassNames=\
-        MutableDataSet
