--- workspaces/liferay-learn-workspace/client-extensions/learn-dxp-importer/src/main/resources/com/liferay/learn/dxp/importer/dependencies/token.properties
@@ -1,12 +1,0 @@
-LIFERAY_LEARN_COMMERCE_DOCKER_IMAGE_TOKEN=[$LIFERAY_LEARN_COMMERCE_DOCKER_IMAGE$]
-LIFERAY_LEARN_COMMERCE_DOCKER_IMAGE_TOKEN_VALUE=liferay/commerce:2.0.5
-LIFERAY_LEARN_COMMERCE_GIT_TAG_TOKEN=[$LIFERAY_LEARN_COMMERCE_GIT_TAG$]
-LIFERAY_LEARN_COMMERCE_GIT_TAG_TOKEN_VALUE=2.0.5
-LIFERAY_LEARN_DXP_DOCKER_IMAGE_TOKEN=[$LIFERAY_LEARN_DXP_DOCKER_IMAGE$]
-LIFERAY_LEARN_DXP_DOCKER_IMAGE_TOKEN_VALUE=liferay/dxp:7.4.13-u29
-LIFERAY_LEARN_PORTAL_DOCKER_IMAGE_TOKEN=[$LIFERAY_LEARN_PORTAL_DOCKER_IMAGE$]
-LIFERAY_LEARN_PORTAL_DOCKER_IMAGE_TOKEN_VALUE=liferay/portal:7.4.3.29-ga29
-LIFERAY_LEARN_PORTAL_GIT_TAG_TOKEN=[$LIFERAY_LEARN_PORTAL_GIT_TAG$]
-LIFERAY_LEARN_PORTAL_GIT_TAG_TOKEN_VALUE=7.4.3.29-ga29
-LIFERAY_LEARN_PORTAL_WORKSPACE_TOKEN=[$LIFERAY_LEARN_PORTAL_WORKSPACE$]
-LIFERAY_LEARN_PORTAL_WORKSPACE_TOKEN_VALUE=portal-7.4-ga29
